Address review comments and fix failing tests
